dispose audit sinks
